+++ packages/kit/types/index.d.ts
@@ -52,7 +52,6 @@ export interface Builder {
 	 */
 	writeClient(dest: string): string[];
 	/**
-	 *
 	 * @param dest
 	 */
 	writePrerendered(